+++ frameworks/vue/src/components/accordion/accordion-item.tsx
@@ -1,9 +1,9 @@
 import type { ItemProps } from '@zag-js/accordion'
 import { type PropType, computed, defineComponent } from 'vue'
+import { Collapsible } from '../..'
 import type { Assign } from '../../types'
 import { useRenderStrategyContext } from '../../utils/render-strategy'
-import { type HTMLArkProps, ark } from '../factory'
-import { PresenceProvider, usePresence } from '../presence'
+import type { HTMLArkProps } from '../factory'
 import { emits } from '../presence/presence.props'
 import { useAccordionContext } from './use-accordion-context'
 import { AccordionItemProvider } from './use-accordion-item-context'
@@ -12,25 +12,26 @@ import { AccordionItemPropsProvider } from './use-accordion-item-props-context'
 export interface AccordionItemProps extends Assign<HTMLArkProps<'div'>, ItemProps> {}
 
 export const AccordionItem = defineComponent<AccordionItemProps>(
-  (props, { slots, attrs, emit }) => {
+  (props, { slots, attrs }) => {
     const accordion = useAccordionContext()
-    const itemState = computed(() => accordion.value.getItemState(props))
-
+    const item = computed(() => accordion.value.getItemState(props))
+    const itemContentProps = computed(() => accordion.value.getItemContentProps(props))
     const renderStrategyProps = useRenderStrategyContext()
-    const usePresenceProps = computed(() => ({
-      ...renderStrategyProps.value,
-      present: itemState.value.isOpen,
-    }))
-    const presenceApi = usePresence(usePresenceProps, emit)
 
-    AccordionItemProvider(itemState)
+    AccordionItemProvider(item)
     AccordionItemPropsProvider(props)
-    PresenceProvider(presenceApi)
 
     return () => (
-      <ark.div {...accordion.value.getItemProps(props)} {...attrs}>
+      // @ts-expect-error TODO fix dir typing
+      <Collapsible.Root
+        open={item.value.isOpen}
+        ids={{ content: itemContentProps.value.id }}
+        {...accordion.value.getItemProps(props)}
+        {...renderStrategyProps.value}
+        {...attrs}
+      >
         {slots.default?.()}
-      </ark.div>
+      </Collapsible.Root>
     )
   },
   {