+++ frameworks/vue/src/components/accordion/accordion-root.tsx
@@ -1,8 +1,11 @@
 import { computed, defineComponent } from 'vue'
 import type { Assign } from '../../types'
-import { type RenderStrategyProps, RenderStrategyProvider } from '../../utils/render-strategy'
+import {
+  type RenderStrategyProps,
+  RenderStrategyProvider,
+  renderStrategyProps,
+} from '../../utils/render-strategy'
 import { type HTMLArkProps, ark } from '../factory'
-import { emits as presenceEmits, props as presenceProps } from '../presence/presence.props'
 import { emits, props } from './accordion.props'
 import { type UseAccordionProps, useAccordion } from './use-accordion'
 import { AccordionProvider } from './use-accordion-context'
@@ -32,8 +35,8 @@ export const AccordionRoot = defineComponent<AccordionRootProps>(
     name: 'AccordionRoot',
     props: {
       ...props,
-      ...presenceProps,
+      ...renderStrategyProps,
     },
-    emits: [...emits, ...presenceEmits],
+    emits,
   },
 )