+++ frameworks/vue/src/components/accordion/examples/basic.vue
@@ -3,19 +3,19 @@ import { ref } from 'vue'
 import { Accordion } from '../..'
 import { ChevronDownIcon } from './icons'
 
-const items = ref(['React', 'Solid', 'Vue'])
+const items = ref(['React'])
 </script>
 
 <template>
-  <Accordion.Root>
+  <Accordion.Root lazyMount>
     <Accordion.Item v-for="item in items" :key="item" :value="item">
       <Accordion.ItemTrigger>
         What is {{ item }}?
-        <Accordion.ItemIndicator><ChevronDownIcon /></Accordion.ItemIndicator>
+        <Accordion.ItemIndicator>
+          <ChevronDownIcon />
+        </Accordion.ItemIndicator>
       </Accordion.ItemTrigger>
-      <Accordion.ItemContent
-        >{{ item }} is a JavaScript library for building user interfaces.</Accordion.ItemContent
-      >
+      <Accordion.ItemContent>{{ item }} is a JavaScript library for building user interfaces.</Accordion.ItemContent>
     </Accordion.Item>
   </Accordion.Root>
 </template>