+++ frameworks/vue/CHANGELOG.md
@@ -6,6 +6,10 @@ description: All notable changes to this project will be documented in this file
 
 ## [Unreleased]
 
+### Added
+
+- Integrate `Collapsible` component with `Accordion` component
+
 ## [1.0.0-0] - 2024-04-19
 
 ### Added